Stream explore page: split into shell + async content component
Extract heavy data fetching (parseSearchFilters + searchJobs) into an
ExploreContent async server component wrapped in Suspense. The page
function now resolves instantly (~5ms) and streams a skeleton fallback
while the DB queries complete (~80-250ms).

- page.tsx: thin shell that reads params/headers and returns Suspense
- explore-content.tsx: async server component with all data fetching
- explore-skeleton.tsx: shared skeleton used by both Suspense fallback
  and loading.tsx (route transitions)
